Refactoring to support Pipelines
